requirements.txt
moviepy  
Pillow==9.4.0